Legerdemain/requirements.txt

streamlit
xhtml2pdf
jinja2
faker
numpy
python-dateutil
